+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
-## [0.15.9] - in progress
+## [0.15.9] - 20204-09-28
 
 
 - Bug fix in `IterableQueue` when using a multiprocessing SimpleQueue.